# Avoid method call if `@transaction_state` is not finalized