@shioyama 2019-04-21
Wow great! `user.id` also faster, that's great. Fantastic! Thanks for responding so quickly.

Just so the PRs are linked: this is in response to #35987 [comment](https://github.com/rails/rails/pull/35987#issuecomment-485222483).