@shioyama 2019-04-21
Just to confirm: I also see a performance bump of about 6-8% or so.