@kaspth 2019-04-21
What would the perf hit be if we abstracted `sync_with_transaction_state if @transaction_state&.finalized?` to something like `sync_with_finalized_transaction_state`?