@shioyama 2019-04-21
@kaspth I think the point is that *any* method call takes a performance hit, so checking an instance variable first (which in 99% of cases is falsey) saves that hit and improves performance.

It's not pretty though, so maybe in the future someone could come up with a cleaner solution with the same performance.