@rafaelfranca 2019-04-22
The code is now harder to understand and maintain. It now open the possibility to someone to introduce a call to `sync_with_transaction_state` that doesn't check by the `@transaction_state` as if this was a valid call. It is not.

I totally understand that this is faster, but it is only faster because the MRI is slow. I don't think we should be spreading conditionals and avoiding method calls breaking encapsulation, increasing maintainability, cost to compensate the limitations of the Ruby VM. If the MRI is slow we improve it, not ask everyone to change their code because we can't make MRI faster.