@shioyama 2019-04-22
#36052 is similar. ivar access over method calls gives a pretty sizeable boost on getters.

Maybe the change could be limited to only `_read_attribute`, since that's the one that originally incurred the 10% hit to `user.name`? The others probably make less of a difference.