@SamSaffron 2019-04-22
@rafaelfranca @kamipo @k0kubun  I wonder if we can rely longer term on methods being inlined / erased with:

```
sync_with_transaction_state 

def sync_with_transaction_state
    sync_with_transaction_state_impl if @transaction_state&.finalized?
end

def sync_with_transaction_state_impl
   .....
end
```

Or even deeper magic...I totally hear @rafaelfranca's concern about risk of regressions due to code being spread out. 

But also feel that the 10% figure is a pretty material win here we should not be giving up.