@kamipo 2019-04-22
Note that `sync_with_transaction_state` without `@transaction_state&.finalized?` is still valid call.
This guard condition is just for early return, if `sync_with_transaction_state` is called when `@transaction_state` isn't finalized, all `fully_committed?`, `committed?`, `fully_rolledback?`, `rolledback?` states are falsy.

https://github.com/rails/rails/blob/1b2efe5a111941d57edbaf2be7f3a94eb02b5f9d/activerecord/lib/active_record/transactions.rb#L480-L492

I agree that it is not pretty. One possible solution making code simple is reverting the GC friendly transaction #9068.
Reverting #9068 completely removes all `sync_with_transaction_state` in our code base.