@SamSaffron 2019-04-22
We need to re-validate if we still need #9068 in current rubies? Do we have a gist demonstrating the pathological cases where it is even needed?