@rafaelfranca 2019-04-23
> But also feel that the 10% figure is a pretty material win here we should not be giving up.

Yes, this is why I didn't reverted this change. The code is now less clear but 10% is a lot, I commented more because I don't want this trend to continue to shave off 1% in micro benchmarks. If we have more low-hanging fruits that yields gains similar to this one I'm fine on applying the same technique.