@kaspth 2019-04-27
@kamipo I'd be curious to see a version that reverted #9068 and what the implications of removing it might be. I remember Jon Wang worked on running multiple Rails applications in a single process back for GSoC 2013, so this could be related to that.